+++ react/TextAreaBlock.tsx
@@ -59,6 +59,7 @@ const TextAreaBlock: FunctionComponent<
   })
 
   const [refidLoading, setRefIdLoading] = useState<any>()
+  const [loading, setLoading] = useState<boolean>(false)
   const [isModalOpen, setIsModelOpen] = useState<boolean>(false)
 
   const { textAreaValue, reviewItems, reviewState } = state
@@ -200,6 +201,7 @@ const TextAreaBlock: FunctionComponent<
   }
 
   const parseText = async () => {
+    setLoading(true)
     const items: any =
       (await ParseText(
         textAreaValue,
@@ -207,6 +209,7 @@ const TextAreaBlock: FunctionComponent<
         customerNumber,
         targetSystem
       )) || []
+    setLoading(false)
 
     const error = !!items.filter((item: any) => {
       return item.error !== null
@@ -355,6 +358,7 @@ const TextAreaBlock: FunctionComponent<
                 <Button
                   variation="secondary"
                   size="regular"
+                  isLoading={loading}
                   onClick={() => {
                     parseText()
                   }}