+++ react/UploadBlock.tsx
@@ -62,6 +62,7 @@ const UploadBlock: FunctionComponent<
   })
 
   const [refidLoading, setRefIdLoading] = useState<any>()
+  const [loading, setLoading] = useState<boolean>(false)
   const [isModalOpen, setIsModelOpen] = useState<boolean>(false)
   const { reviewItems, reviewState } = state
   const apolloClient = useApolloClient()
@@ -166,6 +167,7 @@ const UploadBlock: FunctionComponent<
   }
 
   const parseText = async () => {
+    setLoading(true)
     let textAreaValue = ''
 
     productsArray.forEach((element: any) => {
@@ -179,6 +181,7 @@ const UploadBlock: FunctionComponent<
         customerNumber,
         targetSystem
       )) || []
+    setLoading(false)
 
     const error = !!items.filter((item: any) => {
       return item.error !== null
@@ -463,6 +466,7 @@ const UploadBlock: FunctionComponent<
                 <Button
                   variation="secondary"
                   size="regular"
+                  isLoading={loading}
                   onClick={() => {
                     parseText()
                   }}